Add available field to listing model with default value of true
NOTE: I'm noticing for the first time that when the tests fail, I get a
mongoerror topology destroyed error. Not sure if this commit introduces
it, or what is actually causing it. The tests are fine when they pass
but not very helpful when they fail. The first test to fail will be the
one that is actually failing, and all the other failing ones afterwards
don't matter.